# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CK Empire Builder API
  description: |
    # CK Empire Builder - Advanced Digital Empire Management API
    
    ## Overview
    
    CK Empire Builder is a comprehensive digital empire management platform that provides:
    
    * **Project Management**: Organize and track digital projects
    * **Revenue Tracking**: Monitor income streams and financial metrics
    * **AI Integration**: Advanced AI features for content generation
    * **Ethics Monitoring**: Built-in ethical AI monitoring
    * **Performance Analytics**: Real-time metrics and insights
    * **Cloud Integration**: AWS/GCP integration with auto-scaling
    * **Backup & Recovery**: Automated backup and disaster recovery
    
    ## Features
    
    ### 🔥 Core Features
    - **Project Management**: Create, track, and manage digital projects
    - **Revenue Analytics**: Monitor income streams and financial performance
    - **AI-Powered Content**: Generate content using advanced AI models
    - **Ethics Compliance**: Built-in ethical AI monitoring and compliance
    - **Performance Monitoring**: Real-time analytics and insights
    
    ### ☁️ Cloud Features
    - **AWS Integration**: S3 backup, RDS database, ElastiCache Redis
    - **GCP Integration**: Cloud Storage, Cloud SQL, Memorystore
    - **Auto-scaling**: Kubernetes HPA for automatic scaling
    - **Backup & Recovery**: Automated database backups and restoration
    - **Monitoring**: CloudWatch metrics and alerts
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/ethics/check:
    post:
      summary: Check Content Ethics
      description: Check content for ethical concerns
      tags:
        - Ethics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EthicsCheckRequest'
      responses:
        '200':
          description: Ethics check completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EthicsCheckResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EthicsCheckRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Content to check for ethics
          example: "Content text to analyze..."
        content_type:
          type: string
          enum: [text, image, video]
          default: text
          example: "text"
        user_id:
          type: string
          description: User ID for tracking
          example: "user_123456789"
    EthicsCheckResponse:
      type: object
      properties:
        is_ethical:
          type: boolean
          example: true
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.95
        flagged_issues:
          type: array
          items:
            type: string
          example: []
        recommendations:
          type: array
          items:
            type: string
          example: ["Content appears to be ethical and compliant"]
...